requirements.txt
build
cv-bridge
matplotlib==3.1.2
numpy
opencv-python
osrf-pycommon
Pillow
pyproj
pymap3d
python-qt-binding
PyQt5
PyYAML
rasterio
rospkg
rospy
scipy
setuptools
tf
tk